+++ LiquidacionPeajesNew.Domain/Interfaces/IZonaGaritaRepository.cs
@@ -5,10 +5,10 @@ namespace LiquidacionPeajesNew.Domain.Interfaces
     public interface IZonaGaritaRepository
     {
         Task<IEnumerable<ZonaGaritaEntity>> GetAllAsync();
-        Task<ZonaGaritaEntity> GetByIdAsync(byte id);
-        Task<bool> ExistsAsync(byte id, string name);
+        Task<ZonaGaritaEntity> GetByIdAsync(short id);
+        Task<bool> ExistsAsync(short id, string name);
         Task AddAsync(ZonaGaritaEntity entity);
         Task UpdateAsync(ZonaGaritaEntity entity);
-        Task DeleteAsync(byte id);
+        Task DeleteAsync(short id);
     }
 }
